refactor: Extract SharingBanner from public toolbars
toolbars for share from cozy to cozy and by link
